requirements.txt
python-telegram-bot==20.0
PyMuPDF==1.23.26
python-docx==1.1.0
Pillow==10.2.0
python-dotenv==1.0.1
telegram
python-telegram-bot
python-telegram-bot
reportlab
reportlab
docx2pdf
pdf2docx
PyPDF2
docx2pdf
pdf2docx
PyPDF2
docx2pdf
pdf2docx
Pillow
PyMuPDF
PyPDF2
python-docx
python-dotenv
python-telegram-bot
reportlab
